requirements.txt
uvicorn[standard]
fastapi
google-cloud-speech